+++ app/Http/Controllers/Admin/CourseTopicController.php
@@ -18,18 +18,24 @@ class CourseTopicController extends Controller
         return view('admin.pages.course_topic.create');
     }
 
-    public function customCourseTopic($course_category, $course)
-    {
-        dd($course_category);
-    }
+    // public function customCourseTopic($course_category, $course)
+    // {
+    //     dd($course_category);
+    // }
 
     public function edit(CourseTopic $course_topic)
     {
         return view('admin.pages.course_topic.edit', compact('course_topic'));
     }
 
-    public function destroy(CourseTopic $course_topic)
+    public function destroy($course_topic)
     {
+        $course_topic = CourseTopic::where('slug', $course_topic)->with(['exams'])->firstOrFail();
+
+        if(count($course_topic->exams) > 0){
+            return redirect()->route('course-topic.index')->with('failed', 'island cannot be deleted. It has exams associated to it.');
+        }
+
         if(file_exists(public_path($course_topic->zero_star_island_image))){
             unlink(public_path($course_topic->zero_star_island_image));
         }
